Update nix package checkout and fix broken build on MacOS
Currently node 20 can not be build on arm macos, we use it as a
dependency for neovim (for plugins) and pyright (ansible). To work
around the issue for neovim we remove the `withNodeJs` home-manager
attribute and install nodejs globally. For pyright we need a global
overlay that replaces node 20 with 22.

https://www.github.com/NixOS/nixpkgs/issues/402079